# api.yaml
openapi: 3.0.0
info:
  title: Personal-Trainer-API
  version: 1.0.0
servers:
  - url: /api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    BaseResponse:
      type: object
      required: [status, message, code]
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
        code:
          type: string
          description: Machine-readable response code (e.g., OK, BAD_REQUEST, NOT_FOUND)
    SuccessResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              nullable: true
              description: Any JSON value (object/array/string/number/etc)
              x-go-type: interface{}
            meta:
              nullable: true
              description: Any JSON value (usually object)
              x-go-type: interface{}
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            errors:
              type: array
              nullable: true
              items:
                $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: The field that failed validation
        message:
          type: string
          description: Human-readable error message for this field
...
    CancelBookingRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          enum:
            - schedule_conflict
            - feeling_unwell
            - personal_emergency
            - trainer_request
            - client_request
            - other
          description: Reason for cancellation
        notes:
          type: string
          maxLength: 200
          description: Optional additional details about cancellation
...
    CancelBookingResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              type: object
              required:
                [
                  booking_id,
                  status,
                  cancelled_at,
                  refund_amount,
                  refund_reason,
                  notification_sent,
                ]
              properties:
                booking_id:
                  type: string
                  format: uuid
                status:
                  type: string
                  example: cancelled
                cancelled_at:
                  type: string
                  format: date-time
                refund_amount:
                  type: integer
                  description: Number of session credits refunded (0 if forfeited within 12 hours)
                  example: 1
                refund_reason:
                  type: string
                  enum: [within_12_hours, after_12_hours]
                  description: Why the refund amount was calculated this way
                notification_sent:
                  type: boolean
                  description: Whether cancellation notification was sent to both parties
...
security:
  - bearerAuth: []
paths:
...
  /bookings/{id}/cancel:
    put:
      operationId: CancelBooking
      tags:
        - Booking
      summary: Cancel a confirmed booking
      description: |
        Allows a client or trainer to cancel a confirmed booking (PUT /bookings/:id/cancel).
        Enforces cancellation policy based on time until session:
        - Within 12 hours: No credit refund (credits forfeited)
        - After 12 hours: Full credit refund
        Releases the slot back to trainer's availability and notifies both parties.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Booking ID to cancel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelBookingRequest"
      responses:
        "200":
          description: Booking cancelled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelBookingResponse"
        "400":
          description: Invalid input or validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (user is not the client or trainer of this booking)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict (booking already cancelled or session already started)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /sessions/{id}:
    get:
      security:
        - bearerAuth: []
      operationId: HandleGetSessionById
      summary: Takes the Id of a session and returns data concerning the session
      tags:
        - Booking Session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session fetched successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                status: success
                message: Session Fetched
                code: OK
                data:
                  id: "11111111-1111-1111-1111-111111111111"
                  actual_start: "2026-05-07T10:00:00Z"
                  actual_end: "2026-06-07T10:00:00Z"
                  trainer_joined: false
                  client_joined: false
                  status: booked
                  trainer_note: "Great session and clear communication."
                  created_at: "2026-05-07T10:00:00Z"
        "404":
          description: session with provided ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: booking session not found.
                code: BAD_REQUEST
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: Internal server error
                code: SERVER_ERROR
  /sessions/{id}/start:
    put:
      security:
        - bearerAuth: []
      operationId: HandleStartSession
      summary: A trainer starts a session via this endpoint.
      tags:
        - Booking Session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session started successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                status: success
                message: Session started
                code: OK
                data:
                  id: "11111111-1111-1111-1111-111111111111"
                  actual_start: "2026-05-07T10:00:00Z"
                  actual_end: "2026-06-07T10:00:00Z"
                  trainer_joined: true
                  client_joined: false
                  status: started
                  trainer_note: ""
                  created_at: "2026-05-07T10:00:00Z"
        "400":
          description: Invalid or missing parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: Invalid or missing parameters
                code: BAD_REQUEST
        "404":
          description: session with provided ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: booking session not found.
                code: BAD_REQUEST
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: Internal server error
                code: SERVER_ERROR
  /sessions/{id}/join:
    put:
      security:
        - bearerAuth: []
      operationId: HandleJoinSession
      summary: A client joined a session via this endpoint.
      tags:
        - Booking Session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session joined successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                status: success
                message: Session joined
                code: OK
                data:
                  id: "11111111-1111-1111-1111-111111111111"
                  actual_start: "2026-05-07T10:00:00Z"
                  actual_end: "2026-06-07T10:00:00Z"
                  trainer_joined: true
                  client_joined: true
                  status: in-session
                  trainer_note: ""
                  created_at: "2026-05-07T10:00:00Z"
        "400":
          description: Invalid or missing parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: Invalid or missing parameters
                code: BAD_REQUEST
        "404":
          description: session with provided ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: booking session not found.
                code: BAD_REQUEST
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: Internal server error
                code: SERVER_ERROR
  /sessions/{id}/complete:
    put:
      security:
        - bearerAuth: []
      operationId: HandleCompleteSession
      summary: The session is completed via this endpoint.
      tags:
        - Booking Session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session completed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                status: success
                message: Session completed
                code: OK
                data:
                  id: "11111111-1111-1111-1111-111111111111"
                  actual_start: "2026-05-07T10:00:00Z"
                  actual_end: "2026-06-07T10:00:00Z"
                  trainer_joined: true
                  client_joined: true
                  status: completed
                  trainer_note: ""
                  created_at: "2026-05-07T10:00:00Z"
        "400":
          description: Invalid or missing parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: Invalid or missing parameters
                code: BAD_REQUEST
        "404":
          description: session with provided ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: booking session not found.
                code: BAD_REQUEST
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: Internal server error
                code: SERVER_ERROR
  /sessions/{id}/notes:
    put:
      security:
        - bearerAuth: []
      operationId: HandleTrainersNote
      summary: Trainers submit notes for client.
      tags:
        - Booking Session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [note]
              properties:
                note:
                  type: string
                  format: string
      responses:
        "200":
          description: trainers note sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                status: success
                message: Trainer submitted note
                code: OK
                data:
                  id: "11111111-1111-1111-1111-111111111111"
                  actual_start: "2026-05-07T10:00:00Z"
                  actual_end: "2026-06-07T10:00:00Z"
                  trainer_joined: true
                  client_joined: true
                  status: completed
                  trainer_note: "It was a greaession. Looking forwars for more sessions"
                  created_at: "2026-05-07T10:00:00Z"
        "400":
          description: Invalid or missing parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: Invalid or missing parameters
                code: BAD_REQUEST
        "404":
          description: session with provided ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: booking session not found.
                code: BAD_REQUEST
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: Internal server error
                code: SERVER_ERROR